template version as u64

+++ golem-worker-executor-base/src/durable_host/golem/mod.rs
@@ -652,7 +652,7 @@ impl From<golem_common::model::WorkerMetadata> for golem::api::host::WorkerMetad
             args: value.args,
             env: value.env,
             status: value.last_known_status.status.into(),
-            template_version: value.worker_id.template_version,
+            template_version: value.worker_id.template_version as u64,
             retry_count: 0,
         }
     }
